+++ lib/tremble.dart
@@ -8,6 +8,7 @@ export 'package:tremble/game_area.dart';
 export 'package:tremble/screen_controller.dart';
 export 'package:tremble/second_order_dynamics.dart';
 export 'package:tremble/signal.dart';
+export 'package:tremble/spring.dart';
 export 'package:tremble/sprite.dart';
 export 'package:tremble/sprite_batch.dart';
 export 'package:tremble/state_machine.dart';